Use tab-width as a base for indentation.
`coffee-tab-width` is still used to set the default
indentation. Therefore, the documentation is still correct. This fixes
issue #136 happening when `tab-width` and `coffee-tab-width` where of
different values. Also, we don't need to make `coffee-tab-width` a local
variable since the variable used really is `tab-width`.

This is a bit the reverse of 7976bc. Meantimes, we got b4b2a5 that set
local `tab-width` to the value of `coffee-tab-width` when entering the
mode. The value of `tab-width` could still be altered manually by the
user, by a local-variable block or by some minor modes like
dtrt (indentation detection).